Move partition tests to new class
