+++ Package.swift
@@ -203,7 +203,8 @@ let package = Package(
             sources: [
                 "SwiftObjectServerTests.swift",
                 "SwiftBSONTests.swift",
-                "SwiftCollectionSyncTests.swift"
+                "SwiftCollectionSyncTests.swift",
+                "SwiftObjectServerPartitionTests.swift"
             ],
             swiftSettings: combineFlags()
         ),
@@ -215,7 +216,8 @@ let package = Package(
                 "RLMBSONTests.mm",
                 "RLMObjectServerTests.mm",
                 "RLMWatchTestUtility.m",
-                "RLMCollectionSyncTests.mm"
+                "RLMCollectionSyncTests.mm",
+                "RLMObjectServerPartitionTests.mm"
             ],
             cxxSettings: testCxxSettings
         )
